Remove buggy conditional around related posts header (#901)
This removes a buggy conditional that checks if `related_label` is
available in translations before displaying the header for related
posts but this header already handle missing translation by using a
default string.

With this fix the header is displayed even if the translation for this
label isn't available.

Fix #900