Add `{{ content }}` to `home` layout (#1775)
* Add `{{ content }}` to `home` layout

Allow the author to add content which will appear above the list of
recent posts. This also allows the `home` layout to be extended.

Add text to index.html in tests to indicate that the content appears in
the correct spot.

* Add paragraph markup

Since this is a HTML file, proper paragraph markup is needed.